+++ SPV.CPP/SDK/Implement/SidechainSubWallet.h
@@ -19,7 +19,7 @@ namespace Elastos {
 					const std::string &fromAddress,
 					const nlohmann::json& mainchainAccounts,
 					const nlohmann::json& mainchainAmounts,
-					double fee,
+					uint64_t fee,
 					const std::string &payPassword,
 					const std::string &memo);
